fix(next): correctly bundle next-auth/middleware fixes #6025
